+++ Angular/src/app/dashboard/users/users.component.scss
@@ -1,0 +1,74 @@
+/* Table specific CSS */
+table {
+  width: 100%;
+}
+
+td,
+th {
+  width: 25%;
+}
+
+tr.table-detail-row {
+  height: 0;
+}
+
+tr.table-element-row:not(.table-expanded-row):hover {
+  background: whitesmoke;
+}
+
+tr.table-element-row:not(.table-expanded-row):active {
+  background: #efefef;
+}
+
+.table-element-row td {
+  border-bottom-width: 0;
+}
+
+.table-element-detail {
+  overflow: hidden;
+  display: flex;
+}
+
+.table-element-description {
+  padding: 16px;
+}
+
+.table-element-description-attribution {
+  opacity: 0.5;
+}
+
+/* Beautify the space between expansion panel header, description and icon */
+.mat-expansion-panel-header-title,
+.mat-expansion-panel-header-description {
+  flex-basis: 0;
+  justify-content: space-between;
+  align-items: center;
+  margin-left: 8px;
+}
+
+/* Input form CSS */
+.input-form {
+  min-width: 50%;
+  max-width: 90%;
+  /* width: 100%; */
+}
+
+.input-form-custom {
+  max-width: var(--max-width);
+}
+
+.input-form-small {
+  min-width: 20%;
+  max-width: 20%;
+  /* width: 100%; */
+}
+
+.input-form-half {
+  min-width: 40%;
+  max-width: 100%;
+  /* width: auto; */
+}
+
+.input-form-full-width {
+  width: 90%; /* I like the irony */
+}
